Re-implement SRG overlay
SRG overlays have been broken since the repo was converted from the shorthand
format. This brings the support for overlays/srg_support.xml back to the game.

Overlay is used to generate more complete SRG Mapping table. The mapping tables
provides overview how the SRG requirements are implemented in scope of the
product.