Fix newline discrepancies in jinja macros for file content
On older systems (for example RHEL-7.4) newlines at the end of jinja
macros might get stripped during processing. This commit makes the file
content jinja macros to work in any case by explicitly making sure that
there are no whitespaces at the beginning nor at the end of the variables
which are used in the macros.

This change also requires removal of trailing newlines from test files
in the `tests/shared/audit` as these files are not being trimmed of
whitespaces compared to oval, bash and ansible jinja macros.