SSGTS: profile mode extended to reboot VM before performing the final scan
Some rules (for example `configure_crypto_policy`) require reboot of a machine
after a remediation to fully pass. Therefore, this commit adds reboot functionality
into the libvirt backend of the SSGTS. Reboot is performed only when testing using
the SSGTS profile mode after the remediation of a profile and before the final scan.