ocp4: Remove firewalld checks from profiles
These are not necessary since OCP4 uses iptables. The reasoning behind
using iptables is because it's what the kubelet/kubeproxy uses, and
these components are the ones that are in charge of changing the
firewall configurations. On the other hand, further firwall rules might
be managed by the SDN solution.