The whole /etc/issue file should be evaluated
Added test scenario where the banner is followed by an
extraneous line. This caused the rule to pass unexpectedly.

Updated OVAL check to consider the all lines of /etc/issue the object to
be evaluated and compared against a state.
Also updated Bash remediation to not add extra newline at the end, and
Asnbile remediation to remove any extraneous line in /etc/issue