ocp4: Add initial e2e test
This is an initial and simple implementation of the e2e test for ocp4
compliance.

Currently the test does the following:
* Upload image if needed
* Create namespace
* Deploy compliance operator
* Create ComplianceSuite object
* Check that the ComplianceSuite finishes successfully

Subsequent testing will be done in further iterations. This will
include:

* Applying remediations
* Checking re-scanning to check that the remediations no longer appear
* Doing smoke-test to verify that cluster is still working

You can run the e2e test by doing the following:

```
export PROFILE=moderate
make -f tests/ocp4e2e/Makefile e2e
```

Note that this also updates the compliance-operator version to be
installed to the latest one in the manifest. This helps the installation
go faster (no need to update in the cluster).