Add support for kubernetes remediation type
This introduces support for kubernetes remediation types. They're meant
to be generic kubernetes objects that can be taken into use to remediate
an issue in a Kubernetes distribution. These are basically a super-set
of the `ignition` remediation. `ignition` remediations only deal with
MachineConfig objects, while Kubernetes remediations can do any
kubernetes object. In the future, kubernetes remediations will deprecate
ignition remediations in the compliance-operator and in this content.

Note that this needs https://github.com/OpenSCAP/openscap/pull/1514 to
merge first.