ocp4/utils: Enable build_ds_container.sh to create images in other namespaces
Right now the script is hardcoded to use the `openshift-compliance`
namespace. This makes it so that the aforementioned namespace is taken
as a default, but if you give a second parameter to the script, it'll
use that namespace.

e.g.

    ./utils/build_ds_container.sh ocp4 other-namespace

Will build the content for ocp4, and push the container image to the
namespace called "other-namespace".