Warn about findings from rpm_verify_permissions
There can be cases in which a Profile requires that a file permission be
more strict than package default permissions. In this cases this rule
will report the file changed by the Profile itself as a finding.

Not all permission changes make sense to be incorporated by the
package, and currently there is no mechanism to waive these findings.