Warn about findings from rpm_verify_ownership
There can be cases in which a Profile requires that a file be owned
by root, while the package default owner is a different user.
In these cases this rule will report the change in file ownership
done by the Profile itself as a finding.

Not all ownership changes make sense to be incorporated by the
package, and currently there is no mechanism to waive these
findings.