Fail the build when some of html guides was not generated properly
In order to understand this change, let me please drive your attention to the
cmake macro called `ssg_build_html_guides`. The said macro builds all the HTML
guides for a given product and the guide index. The existence of the guide index
makes the `make` command skip the macro and hence, we need to ensure all the
guides are generated properly, before ever generating the guide index.

Previously, when guide generation failed no error was reported. But more
importantly, subsequent run of the `make` did not generate the guide as the
index was well in place.

Looking at the code, I get a feeling that original authors wanted to make the
build error out on failure (given the exception handling code in the thread
main).