Make OCP4 CPE check more lax
The check was verifying the version contained in the clusteroperators
resource. While this correct, it doesn't hold true in some CI
environments. This resulted in us struggling to do checks in CI...

Talking to the team in charge of building this component, they mentioned
that it's sufficient to check for the existence of this object in order
to verify that the cluster is indeed OpenShift 4.x. So this is what
we're checking now for the cpe.