Remove listing functions from yaml_to_shorthand.py
This patch removes the ability of the `yaml_to_shorthand.py` script to
list inputs and outputs of shorthand generation to its stdout. It's
removing the action argument from the CLI of the script. It leaves the
build action as the only feature of the script. As a result, removing
it simplifies the code in `ssg.build_yaml` module.

Action `list-outputs` only prints the value `--output` command line
argument, therefore it is useless.

Action `list-inputs` lists some of the input files from which the
shorthand.xml will be generated. For example profile files, group files,
and rule files. I find this feature useless, as generating shorthand
involves many other files such as Jinja macro definitions, shared
checks, remediations and others.

The `list-inputs` command is used during configure time in CMake.
However, removing this step from the CMake macro is safe, because all
the files aren't generated, but are part of the source repository
(profile files, group files, and rule files are stored in git).  An
exception is the bash-remediation-functions.xml, but that is already
passed by using `DEPENDS ${BASH_REMEDIATION_FNS_DEPENDS}` in the CMake
command. As the affected CMake command already depends on compiled
profiles, it is sure that the built shorthand.xml file will contain
complete profile data.