Add when clauses for inherited platforms too
Consider the Rule's Group platform while including 'when' clauses to
Ansible snippets.

Some rules have two platforms, a machine platform and a package
platform. One of them is represented of the Rule, and the other is
represented in the Rule's Group.

The platforms are organized like this to due limiation in XCCDF,
multiple platforms in a Rule are ORed, not ANDed.