Replaced XCCDF value instantiation in Bash by a macro call.
The former populate ... mechanism is not Bash, it is a special trick perforemd by our build system.
This trick is confusing, its support in the build system is implemented as a complex code, and
it doesnt support multiple values per remediation intuitively.

This makes the build system involvement explicit, and it opens possibilities to perform implementation
changes without breaking backward compatibility.