Implemented packages: metadata.
Before each test, all scenarios are examined, and list of all packages to install is gathered.
Those packages are installed to the base tests_upladed snapshot
using an SSH command that is determined according to CPEs of the datastream benchmark,
so Fedora ends up with dnf, while RHEL7 with yum.