Augment the etcd_cert_rule rule for ocp4
The check in this patch would inspect the configmap that gets deployed
for the etcd static pod to see whether the expected --cert-file option
is there.

This is a stopgap before we write better OVAL checks that only look at
the master nodes. Checking the manifest on the master nodes itself would
be preferable because it would:
    - check all the master nodes, this patch only checks one configmap
    - we could use more expressive YAML checks to e.g. only check the
      etcd container's command. Because the static pod manifest is a
      value of a configmap key (a yaml in yaml), we just grep the whole
      thing which is fragile.