Generate the cpe-dictionary dynamically
To avoid the need to manually add a new cpe-item for each product,
let's generate the list of cpe-items based on the platforms included in the
Benchmark (i.e. shorthand).

The cpe-item for the derivate product is injected by the
enable_derivatives.py script. The the main product doesn't need to
contain cpe-items related to derivate product anymore.