CIS 1.2.10: Add a rule with disabled OVAL to check for API priority and fairness
Upstream Kubernetes benchmark checks for EventRateLimit which is not
used by OCP. Remove the rule that check for EventRateLimit and instead
add two rules that check that the API priority feature gate is enabled
and that at least the catch-all (that's a reserved name) FlowSchema
object exists. Both rules have their OVAL checks disabled at the moment,
because the Compliance Operator does not have the privileges to read
those objects. A Jira has been filed.