CIS 1.2.16: Ensure that a SCC admission control plugin is set
The vanilla Kubernetes benchmark requires that PCPs are used. OCP does
not use PCPs, but instead use SCCs. This commits adds a rule that checks
that the SCC admission control plugin is set. There are no checks for
details of SCCs, because there are many rules in the 5.2.x section that
do exactly that.