Fix FIPS checks for RHCOS
This adds RHCOS as a FIPS certified OS (it inherits that from RHEL) and
it enables OVAL that's specific to checking RHCOS.

The relevant changes are:

modify sysctl check
-------------------

Given that in the Compliance Operator the OpenSCAP command runs in
offline mode, we can't use the sysctl probe directly. So, instead I
modified that check to use the value from
`/proc/sys/crypto/fips_enabled` directly.

Remove dracut checks
--------------------

Initramfs is managed by ostree, which in turn wraps dracut. It doesn't
use the dracut-fips module directly as that part is handled by ignition
itself. So, these checks have been removed.