OCP4/CIS 3.1.1: Write rule to ensure IdP has been configured
This introduces a rule that makes sure that an IdP has been configured.
Given that it's non-trivial to create an IdP in CI, I took into use the
library that does that for OpenShift's Cluster Authentication Operator
CI. So.... this remediation is done in go-code, as opposed to the
regular format we've been using.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>