ocp4/e2e: Link test failure with sub-tests
Currently, there is a "context" object which contains a lot of the
information the tests need to run. This used to include the test
framework helper instance which one would normally use to output logs or
mark the test as failed.

Unfortunately, this only took into account the main test helper
instance, and so it wasn't entirely clear in what part of the test did
the failure happen. This removes this limitation by passing in the
helper instance along with every function that requires it. This way, we
can pass the sub-test metadata, which will show on what sub-test did the
failure happen.

Note that this PR also increases the manual remediation timeout...
MachineConfigs take a long time to apply unfortunately...

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>