CIS 1.2.35: Add check for api_server_tls_cipher_suites
If the configured cipher suite contains cipher entries outside of the set
defined in the regex, that counts as a fail. This ensures the api server only
uses the modern ciphers, or a subset of, and no extras.

The set is as defined for CIS 1.6.0 benchmark and matches the OCP 4.6 default
(default result PASS).