Update produt based control tests
The tests for product specific rules included Jinja conditionals and a
rule that doesn't exist yet.
But to evaluate product applicability of the rule it needs to be loaded,
so that 'prodtype' can be checked.

Upddate Control.from_control_dict() to handle empty env_yaml and missing
'product' or 'ssg_root'.