Allow product-version-agnostic checks/remediations
This allows product-version-agnostic checks/remediations to be added to
CaC and used ahead of shared content, while still allowing
version-specific content to be favored ahead of both.

That is, given the following:

 - ovals/shared.xml
 - ovals/ubuntu.xml
 - ovals/ubuntu2004.xml

The priority order returned from:
    get_rule_dir_ovals(..., product="ubuntu")
would be:
    ubuntu2004.xml
    ubuntu.xml
    shared.xml

This is useful when OVAL/... content differs between products (e.g.,
RHEL, Ubuntu, ...) but still applies to other versions of the same
product (e.g., 18.04 and 20.04). Duplication thus occurs only at a
product level (if say, Debian chose to reuse Ubuntu content or
visa-versa).

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>