Introduce @ form in all template fields
As noted by @ggbecker in #6659, templates apparently lack the ability to
process @-formed variables in the extra_ovals.yml format. Because of the
way a Rule's yaml gets processed in build_yaml.py, we do not have that
issue due to its preprocessing, though this fix does get applied in both
places.

This allows us to avoid the need to use Jinja processing on
extra_ovals.yml for determining package name, making it look nicer and
be a little more friendly for contributors.

Resolves: #6659

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>