Show which rule failed during template builds
Previously, when a rule failed to build when applying the template
(e.g., using a glob in a package removal template), we wouldn't
necessarily know which rule failed to build. Print an error message
showing which rule this was and then re-throw the error so that the
builder can see which rule caused the build failure.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>