Only convert platform names when products match
A rule can be loaded from yaml in any context, even when building
unrelated products. A rule is completely loaded, and then its prodtype
is assessed against the product being built.

This limits the conversion of platform names to cpe names to situations
where the rule is compatible with the product.