Automatically prodtype rules in a profile
The new autoprodtyper.py utility automatically prodtypes all rules in
a given profile, adding the specified product. This leverages the work
previously done in utils/mod_protype.py to seamlessly edit prodtypes
with minimal diff (modifying only the prodtype line). If we assume a
profile for a product was created with intent (and that the rules in it
were intended to work for the product the profile is in), then
automatically adding this prodtype to the rules saves a bit of effort.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>