Updated checks and remediations of the sshd template.
Configuration of sshd moves from one config file to a config directory.
Therefore, checks should consider all those files, and the remediation should aim
to deliver fixes to one of those files in the config directory.

Tests that interact with this behavior have been added and are applicable for Fedora and RHEL9 products.