Support variable-based CPE platforms
When we generate pseudo-CPEs based off of variables, we load (and then
re-write) the XML of OVAL checks. The rewriting (via ElementTree) causes
an oval: namespace prefix to be added into all platform elements. This
causes later steps (which operate strictly on strings and not on parsed
element trees) to fail.

Update them to use the correct namespace.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>