Make check for chrony vs ntp product independent
With the bash macros introduced in the previous commit, we can make this
check independent of operating system.

This does currently assume that "chrony" and "ntp" are the correct
package names across operating systems, however, we have existing
infrastructure we could reuse at a later point in time if/when packages
with different names are used with this macro.

This macro could also be used in bash fixes for package inventory CPEs,
whereby if a package isn't installed, the fix wouldn't be run (when
running from a straight-profile hardening script instead of from oscap).

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>