OCP4: Check that either the OAuth server or any of the clients have inactivity timeout set
The patch adds two rules (auth_inactivity_timeout and
oauthclient_inactivity_timeout) that check the inactivity timeout for the
OAuth server or the clients respectively. These rules are not included
in any of the profiles, but are referenced from the hand-crafted OVAL of
oauth_or_oauthclient_inactivity_timeout.

The oauth_or_oauthclient_inactivity_timeout rule is included from the
OCP4 moderate profile.

At the moment, the remediation uses a hardcoded timeout of 10 minutes.