Add SCE (Bash) as supported template language
SCE differs from OVAL and other "languages" in that it theoretically
supports multiple backend languages (Bash, Python, Perl, ...). However,
in order for SCE to be effective, the interpreter for said language
needs to present on the target system.

Thus, when generating templates, require the language in addition to the
SCE context. We introduce `sce-bash` in this commit, to handle any
templates wishing to introduce bash SCE content. This allows us to infer
the correct suffix to use for templated SCE content as the build system
maps by language name and not by values inherit in the template itself.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>