Add templated SCE content to SCE metadata
SCE generation works differently from the OVAL check system in that we
build SCE content very early in the build system. This content is two
fold: the actual check content (needed much later in the build, and so
could be built later) and the metadata of which rules actually have SCE
content and what properties they request (necessary early in the build
to generate shorthand).

In adding templated content, we still need its metadata early (and so,
need to load and parse it from build_sce.py) but defer writing its
content out later (until build_templated_content.py is run).

Since this is still in time for DS generation, this is fine.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>