Don't execute fix_rule in dry_run mode
Per feedback from Jan Cerny, don't execute fix_rules (even though it is
a no-op) and thus print the before/after of the rules that need
changing; only print the path that needs changing.

This makes the output easier to read for large change-sets at the
expense of not always making it clear to the caller what needs to
change.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>