Allow missing /etc/motd file for Ubuntu CIS
Ubuntu does not provision an /etc/motd by default. Because most
benchmarks do not mandate a /etc/motd exists (but if it does, that the
permissions are configured correctly), we should add the
missing_file_pass option to the template. If this file is required in
another benchmark, it should be added in a separate rule and when the
audit is re-run, it will pass/fail accordingly.

Internal commit: b12417ca36794a012cbc812c90d7ff6ad3c21e3d

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>