Prioritize install_smartcard_packages like package_*_installed
Several smartcard rules have the problem that install_smartcard_packages
is not appropriately ordered. In #7208, we renamed the rule to match the
pattern of package_*_installed. However, this approach breaks any
existing tailoring and thus wasn't desired. Move to updating the Python
to order this rule as a special case.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>