Only template rules with variables
The audit_rules_privileged_commands_unix2_chkpwd rule lacks variables on
most products, in addition to its limited prodtype. Because the existing
test harness lacks understanding of prodtype (for including rules in the
tarball), we don't check it yet either. This causes issues when the
template is present but has empty variables on the particular product
(such as when this rule is executed under Ubuntu)

Ultimately we should probably check prodtype in the future, but it
outside the scope of this PR.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>