OCP4: The moderate profile requires the audit profile to be WriteRequestBodies
This is needed to cover AU-3(1).

Unfortunately, our CI infra is not yet equipped to handle tests that
need different results for different benchmarks and in this case it
would be needed - CIS requires the Default audit profile to be used, so
the rule passes by default and doesn't need a remediation whereas
moderate checks for WriteRequestBodies so the rule fails by default and
needs to be remediated. Therefore the e2e test is removed for this rule
until we enhance the e2e test machinery.