Fix check and remediations of auditd_overflow_action.
The check was generating a new input to the auditd.conf file and without
spaces between the separator (equal sign). This caused auditd failing to
start since it's mandatory to have a space between the separator. It
also introduces case insensitivity for the check since the paramaters
and values are case insensitive.