OCP4: Don't use variables for kubelet threshold checks
This removes the use of variables from the kubelet threshold checks.
While the idea of using variables is good, at the moment, variables are
not really exposed and thus visible well through the compliance
operator, which means that the admins don't really know what to do to
make the rules pass.

Let's just check for any value in the kubelet threshold checks for the
moment.

I think long term the variables are a good idea and hopefully we'll be
able to get back to using them in the future once we merge the
remediation templating, which is why I haven't removed the variable
definitions in the ".var" files completely yet.