Add script to get statistics out of a control file
This is an initial implementation for getting stats out of a control
file. The script will take an id and a level, and output some relevant
numbers out of that. Here's a sample output

```
./utils/controleval.py stats -i cis_rhel7 -l l2_server
Total controls = 248
Applicable = 100.0% -- 248 / 248
Assessed = 68.55% -- 170 / 248

Automated = 60.48% -- 150 / 248
Supported = 0.0% -- 0 / 248
Documentation = 0.0% -- 0 / 248
Inherently Met = 0.0% -- 0 / 248
Partial = 8.06% -- 20 / 248
```

The idea is to give a notion of how far are we into finishing evaluating
a certain benchmark.

More enhancements can be done in the future, such as outputting which
controls are missing or outputting these categories in such a way that
they could be ingested by a UI.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>