Fix `render-policy.py` to use `products/` directory
When the `products/` directory was introduced, the change wasn't
reflected in some of the libraries for the utilities in the project. one
example was the `render-policy.py`, for which the underlying library
still looked for products on the top level directory of the repo. This
fixes that, thus fixing the utility.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>