Add directory to add tailorings for OCP
Now that we're introducing remediations that require values to be set,
we need to be able to pass those values somehow. Normally this is done
via a tailoring... so let's add a directory where we can do that. The
e2e framework needs to detect this directory and use the tailoring
insead when needed.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>