OCP4: Add rule to check that the default policy is to reject images
Container runtimes have the capability of specifying what images to
accept and what policy to use for each registry. The intent is to ensure
that integrity verification is enabled for relevant repos.

We, thus, create a rule to ensure that the default policy is to reject,
and that administrators explicitly enable each repo they're actually
interested in using. This addresses several NIST controls which are
linked on the rule.

The remediation introduces a policy that ensures that Red Hat registries
are signature-verified, but that is lenient towards the images we use
for testing (e.g. coming from OpenShift CI). We cannot create an
automated remediation here as this configuration is very
deployment-specific and we cannot guess what registries are customers
using, the signature type and policy they'd like to enforce. However, at
least we can get them thinking.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>