OCP4 & RHCOS4: Migrate OpenControl evaluations to controls structure
Using automation, this migrates the existing evaluations done in
OpenControl (from https://github.com/ComplianceAsCode/redhat ) towards
the control structure.

This introduces the nist controls and its three baselines (low, moderate
and high), and so, each was marked appropriately.

Note that OCP4 is the one with the most complete evaluation, as RHCOS4
was handled using a combination of gdocs and excel sheets as opposed to
using OpenControl directly.

This also modifies the profiles to take this into use.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>