Remove file-level dependencies
In this patch, we make custom targets depending only on other targets,
not on files.  In other words, we remove file-level dependencies from
our targets.

File-level dependencies mean that the target depends on a file and is
rebuilt each time the file is changed.

We are aware that file-level dependencies is a CMake best practice that
supports incremental rebuilds. However, we currently discourage
incremental rebuilds in our project and they are broken. We also don't
expect that we will ever fix incremental rebuilds, because that would
mean to track all the source files in CMakeLists and to move the logic
from our Python build scripts to CMake, which isn't feasible.

If we decide to not support incremental rebuilds in our project we can
remove file-level dependencies, It would help us to simplify the
dependencies and track them.