Compare OCIL checks in compare_ds.py
OCIL checks will be compared between the new and the old data stream.
Namely, a diff of the contents of the question_text element will be
shown.

Note that some of the OVAL-related code has been generalized to reuse
the code because the OCIL checks are referenced in a similar way as OVAL
checks in the XCCDF benchmark.