Remove support for multiple directories
Currently the combine_remediations.py script supports to pass multiple
directories as `fix_dirs` positional arguments. However, the
combine_remediations.py is called only at a single place where only a
single directory is provided as an argument. Therefore, we don't need to
have the script to support multiple positional arguments. Instead, we
can rename the argument to better describe its purpose and change it to
an option. As can be seen from the single call of
combine_remediations.py in CMake, this argument contains a path to the
directory where the remediations generated by the templating engine are
stored: `build/${PRODUCT}/fixes_from_templates/${LANGUAGE}` So we think
that a better name for the option will be `--fixes-from-templates-dir`.