Remove the language part of the directory path
The combine_remediation.py is called only at a single place.
We know that the last part of the directory path in the
--output-dir and --fixes-from-templates-dir is the language and
has the same value as the value of --remediation-type argument
which is a required argument. We can construct the path dynamically
inside the script instead of having it hard-coded in CMake.