Remove ident expanstion via XSLT
We can do the same thing while outputting the shorthand.xml.
The shorthand is now generated with the final form of ident CCEs.

The processing of idents was updated to not handle product specific
idents, as the compiled rules contain only the relevant idents.
The project only has CCE identifiers, so the SITG ident handling was
removed, along with the custom-cce-uri.