Fix broken test scenario
During review of #7750 we have discovered that for this test scenario
the rule fails not because the `SELINUX=permissive` but because
no `SELINUX=` entry exists. The missing `SELINUX=` entry was caused
by a typo in the sed expression which effectively replaced
`SELINUX=enforcing` by just `permissive`. As a result, oscap didn't
find the OVAL object `oval:ssg-object_etc_selinux_config:obj:1`.