Remove OpenStack products and checks
These have not been maintained for a long time and don't really work.

Let's remove them to avoid further confusion.

If/When OpenStack checks get created in CaC, they'll need to recreate
the `applications/openstack` directory and a product for the relevant
version (e.g. the current RHOSP version is `16.1` and work is on-going
for 17).

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>