Change oVirt CPE to identify both Host and Manager
This updates and renames the CPE oVirt platforms to identify systems
acting as a Host or Manager. (ovirt)
And aplatform to identify systems with no oVirt role.
The rule warnings are updated to reflect its applicability on Hosts and
Managers.
This also remove dpkg tests, as they are not needed.