Ignore empty XCCDFEntity non-mandatory keys
Empty keys where being added as None, but this can mess up loading of
content where a list is expected.

Not adding empty non-mandatory keys allows the keys to be empty in the
rule.yml, making it possible to have cleaner rules when macros
ovirt_rule_notapplicable_warning() and rule_notapplicable_when_ovirt_installed()
are used.