Determine the platform name for CIS profile descriptions and rationale
We recently added a profile for the EKS CIS benchmark. Luckily, we were
able to re-use many existing rules from the OCP CIS benchmark, but when
you render the rules in an EKS deployment, they reference "OpenShift" in
various descriptions or rationale. This can be confusing to end users
who might not understand the linkage between rules and profiles in this
repository.

This commit defines a variable for rules that are shared between the EKS
and OCP CIS benchmark and displays it appropriately depending on the
product type.