OCP: Only run controller_insecure_port_disabled for OCP 4.6-4.9
Commit:
https://github.com/kubernetes/kubernetes/commit/e481d9996591331d6173e4cb146afcaf5e0308f1

Removed the port option of kube-controller-manager and refactored the
insecure port handling. Instead of the port option, there's now
'secure-port' that defaults to 10257 and the insecure handling is
enabled by setting secure-port to 0. Before we write a rule for that,
let's fix the existing rule to unblock our CI.