Propose a different layout for kubernetes rules
Initially, this repository only contained profiles for OCP. With the
addition of EKS profiles, we now have two distributions and we should
think about how we want to organize the rules for kubernetes platforms.

The new layout must:

  - Render accurate content for each kubernetes distribution
  - Allow for rule re-use wherever possible
  - Keep existing rule names for backwards compatibility

The first point is important so that end-users are presented with
content that matches the official benchmarks word-for-word.

This commit adds a new group under `applications/` for `kubernetes`
distributions. Each rule added to this new layout should be written as
generically as possible if it pertains to more than one kubernetes
platform (e.g., EKS or OCP). Existing rules under
`applications/openshift/` will be ported to `applications/kubernetes/`
without changing the rule name to preserve backwards compatibility.
Those changes will also include parameterization templating in the rule
to support each platform currently using it.

This will result in template-heavy rules, but it provides the following
benefits:

  - All kubernetes-specific rules are in a platform agnostic location
  - We won't have issues with multiple sub-groups with the same name
    (e.g., `applications/kubernetes/eks/logging` and
    `applications/kubernetes/openshift/logging` see issue 8190) which
    breaks the content build process
  - We can still reuse rules across platforms

Additionally, this commit adds a new rule for the EKS CIS profile as an
example. Further, more complicated examples will be ported in future
patches.