Parameterize kubernetes worker rules
Some of the worker rules are used in both EKS and OpenShift profiles.
This commit updates the rules to use parameters based on the product
(either EKS or OCP) so that the descriptions, rationale, and OCIL are
accurate for the user and don't bleed across products.

This is related to issue #8198