Add EKS CIS section 2.1.1 rule
This rule is functionally trivial since it doesn't perform a scan using
the compliance operator and it doesn't remediate an issue, yet. To do
that, we require some additional functionality in compliance-operator to
interact with AWS endpoints.

In the mean time, the rule is informative about how to perform the audit
and how to remediate it.