Generate srgmap tables using Jinja
This commit changes the generator of table-product-srgmap.html and
table-product-srgmap-flat.html files. The current generator implemented
by XSLT is replaced by a Jinja template and a Python script
utils/gen_srg_table.py. Instead of processing XCCDF and OCIL XML files
we will load the data from compiled rules and compiled profiles.  This
change save us approximately 5 seconds during the build of a single
product.  The aim is to generate almost the same table as we generate
currently.

We remove some of the parameters of the CMake macro
ssg_build_html_srgmap_tables because we always call this macro with the
same parameter values. We won't generate xccdf-linked-srg-overlay.xml
because it's an interim file which isn't used or shipped anywhere.