Partially migrate rule to template
Migrate OVAL to template so it is easier to introduce this change, also
this allows an easier future maintenance. DISA's Ol8 STIG allows a few
different confs for disk_full_action. So this makes possible the use of
regex for that, while keeping the setup for other profiles. Introduce
the var value_regex, so a future complete migration is also easier

Signed-off-by: Edgar Aguilar <edgar.aguilar@oracle.com>