Make taylor reg ex to /etc/group and /etc/passwd structure.
The use of * in the regex for the passwd and group file match is a
bit too permisive.

What the regex should more closely match the file format:
passswd:  string:string:int:int:other stuff
group: string:string:int:other stuff

All we really care about in the match is up to otherstuff.